requirements.txt
ezapi-tmdb==0.8.1
plotly==5.13.0
streamlit==1.17.0
